Remove `Tracer` class, move `callSiteSupplierFactory` to `Traces`.
