+++ reactor-core/src/main/java/reactor/core/publisher/SignalLogger.java
@@ -87,7 +87,7 @@ final class SignalLogger<IN> implements SignalPeek<IN> {
 		this.fuseable = source instanceof Fuseable;
 
 		if (correlateStack) {
-			operatorLine = Traces.extractOperatorAssemblyInformation(Tracer.callSiteSupplierFactory.get().get());
+			operatorLine = Traces.extractOperatorAssemblyInformation(Traces.callSiteSupplierFactory.get().get());
 		}
 		else {
 			operatorLine = null;
